requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
fastapi==0.127.0
graphql-core==3.2.7
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lia-web==0.2.3
packaging==25.0
protobuf==6.33.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
starlette==0.50.0
strawberry-graphql==0.287.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
